+++ test/plugins/template.engine/cases/case32.js
@@ -1,5 +1,5 @@
 export const title       = '32) g-bind width dynamic with delay';
-export const description = 'g-bind with $.dynamic(value, duration, delay)';
+export const description = 'g-bind with $$.dynamic(value, duration, delay)';
 
 export function script () {
   const svg = gSVG(document.querySelector('#svg'));
@@ -13,9 +13,9 @@ export function script () {
 }
 
 export default `<svg viewBox="0 0 100 100" id="svg" style="width: 100px; height: 100px">
-  <circle cx="20" cy="20" :r="$.dynamic(20, 1000, time1)"      fill="red"/>  
-  <circle cx="40" cy="40" :r="$.dynamic(20, 1000, time2)" fill="blue"/>  
-  <circle cx="60" cy="60" :r="$.dynamic(20, 1000, time3)" fill="green"/>  
-  <circle cx="80" cy="80" :r="$.dynamic(20, 1000, time4)" fill="orange"/>  
+  <circle cx="20" cy="20" :r="$$.dynamic(20, 1000, time1)"      fill="red"/>  
+  <circle cx="40" cy="40" :r="$$.dynamic(20, 1000, time2)" fill="blue"/>  
+  <circle cx="60" cy="60" :r="$$.dynamic(20, 1000, time3)" fill="green"/>  
+  <circle cx="80" cy="80" :r="$$.dynamic(20, 1000, time4)" fill="orange"/>  
 </svg>
 <pre id="result"></pre>`;